consolidated controller logic for oauth
