+++ src/db/mem.js
@@ -7,6 +7,7 @@ class Mem {
       this.store = {}; // map of key, value
       this.indexes = {}; // map of key, Set
       this.cleanup = {}; // map of indexes for cleanup when store key is deleted
+      this.ttlStore = {}; // map of key to ttl
 
       return this; // Return the newly-created instance
     })();
@@ -37,9 +38,10 @@ class Mem {
 
     this.store[k] = JSON.stringify(val);
 
-    // TODO: ttl expiry
-    // if (ttl) {
-    // }
+    if (ttl) {
+      // TODO: implement cleanup of TTL
+      this.ttlStore[k] = Date.now() + ttl * 1000;
+    }
 
     // no ttl support for secondary indexes
     for (const idx of indexes || []) {
